refactor: move events components into features with stubs
